refactor: remove decprecated codes
