+++ enjoy/src/main/whisper.ts
@@ -152,102 +152,6 @@ class Whipser {
     });
   }
 
-  async transcribeFromAzure(file: string): Promise<Partial<WhisperOutputType>> {
-    const webApi = new Client({
-      baseUrl: process.env.WEB_API_URL || WEB_API_URL,
-      accessToken: settings.getSync("user.accessToken") as string,
-      logger: log.scope("api/client"),
-    });
-    const { token, region } = await webApi.generateSpeechToken();
-    const sdk = new AzureSpeechSdk(token, region);
-
-    const results = await sdk.transcribe({
-      filePath: file,
-    });
-
-    const transcription: TranscriptionResultSegmentType[] = [];
-    results.forEach((result) => {
-      logger.debug(result);
-      const best = take(sortedUniqBy(result.NBest, "Confidence"), 1)[0];
-      const words = best.Display.trim().split(" ");
-
-      best.Words.map((word, index) => {
-        let text = word.Word;
-        if (words.length === best.Words.length) {
-          text = words[index];
-        }
-
-        if (
-          index === best.Words.length - 1 &&
-          !text.trim().match(END_OF_WORD_REGEX)
-        ) {
-          text = text + ".";
-        }
-
-        transcription.push({
-          offsets: {
-            from: word.Offset / 1e4,
-            to: (word.Offset + word.Duration) / 1e4,
-          },
-          timestamps: {
-            from: milisecondsToTimestamp(word.Offset / 1e4),
-            to: milisecondsToTimestamp((word.Offset + word.Duration) * 1e4),
-          },
-          text,
-        });
-      });
-    });
-
-    return {
-      engine: "azure",
-      model: {
-        type: "Azure AI Speech",
-      },
-      transcription,
-    };
-  }
-
-  async transcribeFromCloudflare(
-    file: string
-  ): Promise<Partial<WhisperOutputType>> {
-    logger.debug("transcribing from CloudFlare");
-
-    const data = fs.readFileSync(file);
-    const res: CfWhipserOutputType = (
-      await axios.postForm(`${AI_WORKER_ENDPOINT}/audio/transcriptions`, data, {
-        headers: {
-          Authorization: `Bearer ${settings.getSync("user.accessToken")}`,
-        },
-      })
-    ).data;
-    logger.debug("transcription from Web,", res);
-
-    const transcription: TranscriptionResultSegmentType[] = res.words.map(
-      (word) => {
-        return {
-          offsets: {
-            from: word.start * 1000,
-            to: word.end * 1000,
-          },
-          timestamps: {
-            from: milisecondsToTimestamp(word.start * 1000),
-            to: milisecondsToTimestamp(word.end * 1000),
-          },
-          text: word.word,
-        };
-      }
-    );
-    logger.debug("converted transcription,", transcription);
-
-    return {
-      engine: "cloudflare",
-      model: {
-        type: "@cf/openai/whisper",
-      },
-      transcription,
-    };
-  }
-
   /* Ensure the file is in wav format
    * and 16kHz sample rate
    */
@@ -353,57 +257,6 @@ class Whipser {
         }
       });
     });
-  }
-
-  groupTranscription(
-    transcription: TranscriptionResultSegmentType[]
-  ): TranscriptionResultSegmentGroupType[] {
-    const generateGroup = (group?: TranscriptionResultSegmentType[]) => {
-      if (!group || group.length === 0) return;
-
-      const firstWord = group[0];
-      const lastWord = group[group.length - 1];
-
-      return {
-        offsets: {
-          from: firstWord.offsets.from,
-          to: lastWord.offsets.to,
-        },
-        text: group.map((w) => w.text.trim()).join(" "),
-        timestamps: {
-          from: firstWord.timestamps.from,
-          to: lastWord.timestamps.to,
-        },
-        segments: group,
-      };
-    };
-
-    const groups: TranscriptionResultSegmentGroupType[] = [];
-    let group: TranscriptionResultSegmentType[] = [];
-
-    transcription.forEach((segment) => {
-      const text = segment.text.trim();
-      if (!text) return;
-
-      group.push(segment);
-
-      if (
-        !MAGIC_TOKENS.includes(text) &&
-        segment.text.trim().match(END_OF_WORD_REGEX)
-      ) {
-        // Group a complete sentence;
-        groups.push(generateGroup(group));
-
-        // init a new group
-        group = [];
-      }
-    });
-
-    // Group the last group
-    const lastSentence = generateGroup(group);
-    if (lastSentence) groups.push(lastSentence);
-
-    return groups;
   }
 
   registerIpcHandlers() {
