+++ enjoy/src/types/enjoy-app.d.ts
@@ -253,10 +253,7 @@ type EnjoyAppType = {
   };
   transcriptions: {
     findOrCreate: (params: any) => Promise<TranscriptionType>;
-    process: (params: any, options: any) => Promise<void>;
     update: (id: string, params: any) => Promise<void>;
-    onProgress: (callback: (event, progress: number) => void) => void;
-    removeProgressListeners: () => Promise<void>;
   };
   waveforms: {
     find: (id: string) => Promise<WaveFormDataType>;
